(MISC): extract utils file for typing package
